pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "holographic-cosmo-engine"
version = "0.1.1"
description = "Hybrid AdS5/dS4 holographic cosmology engine (PINN + spectral + analysis)"
readme = "README.md"
requires-python = ">=3.9"
authors = [{name="Hugo Hertault"}]
license = {text = "MIT"}
dependencies = [
  "numpy",
  "scipy",
  "matplotlib",
  "pyyaml",
  "torch",
]

[project.optional-dependencies]
dev = ["pytest","ruff"]

[project.scripts]
hce-train = "hce.cli:main_train"
hce-analyze = "hce.cli:main_analyze"
hce-spectrum = "hce.cli:main_spectrum"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
